Generate OpenAPI for /status routes
