+++ common/cosmwasm-smart-contracts/mixnet-contract/Cargo.toml
@@ -22,6 +22,7 @@ thiserror = { workspace = true }
 contracts-common = { path = "../contracts-common", package = "nym-contracts-common", version = "0.5.0" }
 serde-json-wasm = { workspace = true }
 humantime-serde = { workspace = true }
+utoipa = { workspace = true }
 
 # TO CHECK WHETHER STILL NEEDED:
 log = { workspace = true }
