+++ nym-api/nym-api-requests/src/models.rs
@@ -4,7 +4,7 @@
 use crate::helpers::unix_epoch;
 use crate::nym_nodes::NodeRole;
 use crate::pagination::PaginatedResponse;
-use cosmwasm_std::{Addr, Coin, Decimal};
+use cosmwasm_std::{Addr, Coin, Decimal, Uint128};
 use nym_mixnet_contract_common::families::FamilyHead;
 use nym_mixnet_contract_common::mixnode::MixNodeDetails;
 use nym_mixnet_contract_common::reward_params::{Performance, RewardingParams};
@@ -22,8 +22,7 @@ use std::net::IpAddr;
 use std::ops::{Deref, DerefMut};
 use std::{fmt, time::Duration};
 use time::OffsetDateTime;
-use utoipa::openapi::{schema::SchemaType, Object, ObjectBuilder};
-use utoipa::{ToResponse, ToSchema};
+use utoipa::{IntoParams, ToResponse, ToSchema};
 
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq, JsonSchema)]
 pub struct RequestError {
@@ -85,7 +84,7 @@ impl MixnodeStatus {
     }
 }
 
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
+#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema, ToSchema)]
 #[cfg_attr(feature = "generate-ts", derive(ts_rs::TS))]
 #[cfg_attr(
     feature = "generate-ts",
@@ -96,7 +95,7 @@ pub struct MixnodeCoreStatusResponse {
     pub count: i32,
 }
 
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
+#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema, ToSchema)]
 #[cfg_attr(feature = "generate-ts", derive(ts_rs::TS))]
 #[cfg_attr(
     feature = "generate-ts",
@@ -107,7 +106,7 @@ pub struct GatewayCoreStatusResponse {
     pub count: i32,
 }
 
-#[derive(Clone, Copy, Debug, Serialize, Deserialize, JsonSchema)]
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, JsonSchema, ToSchema)]
 #[cfg_attr(feature = "generate-ts", derive(ts_rs::TS))]
 #[cfg_attr(
     feature = "generate-ts",
@@ -117,19 +116,45 @@ pub struct MixnodeStatusResponse {
     pub status: MixnodeStatus,
 }
 
-#[derive(Clone, Debug, Default, Serialize, Deserialize, JsonSchema)]
+#[derive(Clone, Debug, Default, Serialize, Deserialize, JsonSchema, ToSchema)]
 pub struct NodePerformance {
+    #[schema(value_type = String)]
     pub most_recent: Performance,
+    #[schema(value_type = String)]
     pub last_hour: Performance,
+    #[schema(value_type = String)]
     pub last_24h: Performance,
 }
 
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
+#[derive(ToSchema)]
+#[schema(title = "MixNodeDetails")]
+pub struct MixNodeDetailsSchema {
+    // TODO dz copy from cosmwasm crate
+    /// Basic bond information of this mixnode, such as owner address, original pledge, etc.
+    pub bond_information: String,
+
+    // TODO dz copy from cosmwasm crate
+    /// Details used for computation of rewarding related data.
+    pub rewarding_details: String,
+
+    // TODO dz copy from cosmwasm crate
+    /// Adjustments to the mixnode that are ought to happen during future epoch transitions.
+    pub pending_changes: String,
+}
+
+#[derive(ToSchema)]
+#[schema(as = String, title = Decimal)]
+pub struct DecimalSchema;
+
+#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema, ToSchema)]
 pub struct MixNodeBondAnnotated {
     pub mixnode_details: MixNodeDetails,
+    #[schema(value_type = String)]
     pub stake_saturation: StakeSaturation,
+    #[schema(value_type = String)]
     pub uncapped_stake_saturation: StakeSaturation,
     // NOTE: the performance field is deprecated in favour of node_performance
+    #[schema(value_type = String)]
     pub performance: Performance,
     pub node_performance: NodePerformance,
     pub estimated_operator_apy: Decimal,
@@ -160,7 +185,7 @@ impl MixNodeBondAnnotated {
     }
 }
 
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
+#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema, ToSchema)]
 pub struct GatewayBondAnnotated {
     pub gateway_bond: GatewayBond,
 
@@ -191,13 +216,16 @@ pub struct GatewayDescription {
     // for now only expose what we need. this struct will evolve in the future (or be incorporated into nym-node properly)
 }
 
-#[derive(Debug, Serialize, Deserialize, JsonSchema)]
+#[derive(Debug, Serialize, Deserialize, JsonSchema, ToSchema, IntoParams)]
 pub struct ComputeRewardEstParam {
+    #[schema(value_type = String)]
     pub performance: Option<Performance>,
     pub active_in_rewarded_set: Option<bool>,
     pub pledge_amount: Option<u64>,
     pub total_delegation: Option<u64>,
+    #[schema(value_type = CoinSchema)]
     pub interval_operating_cost: Option<Coin>,
+    #[schema(value_type = String)]
     pub profit_margin_percent: Option<Percent>,
 }
 
@@ -206,7 +234,7 @@ pub struct ComputeRewardEstParam {
     feature = "generate-ts",
     ts(export_to = "ts-packages/types/src/types/rust/RewardEstimationResponse.ts")
 )]
-#[derive(Clone, Copy, Debug, Serialize, Deserialize, JsonSchema)]
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, JsonSchema, ToSchema)]
 pub struct RewardEstimationResponse {
     pub estimation: RewardEstimate,
     pub reward_params: RewardingParams,
@@ -215,15 +243,16 @@ pub struct RewardEstimationResponse {
     pub as_at: i64,
 }
 
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
+#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema, ToSchema)]
 pub struct UptimeResponse {
+    #[schema(value_type = u32)]
     pub mix_id: MixId,
     // The same as node_performance.last_24h. Legacy
     pub avg_uptime: u8,
     pub node_performance: NodePerformance,
 }
 
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
+#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema, ToSchema)]
 pub struct GatewayUptimeResponse {
     pub identity: String,
     // The same as node_performance.last_24h. Legacy
@@ -231,7 +260,7 @@ pub struct GatewayUptimeResponse {
     pub node_performance: NodePerformance,
 }
 
-#[derive(Clone, Copy, Debug, Serialize, Deserialize, JsonSchema)]
+#[derive(Clone, Copy, Debug, Serialize, Deserialize, JsonSchema, ToSchema)]
 #[cfg_attr(feature = "generate-ts", derive(ts_rs::TS))]
 #[cfg_attr(
     feature = "generate-ts",
@@ -239,16 +268,18 @@ pub struct GatewayUptimeResponse {
 )]
 pub struct StakeSaturationResponse {
     #[cfg_attr(feature = "generate-ts", ts(type = "string"))]
+    #[schema(value_type = String)]
     pub saturation: StakeSaturation,
 
     #[cfg_attr(feature = "generate-ts", ts(type = "string"))]
+    #[schema(value_type = String)]
     pub uncapped_saturation: StakeSaturation,
     pub as_at: i64,
 }
 
 pub type StakeSaturation = Decimal;
 
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
+#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema, ToSchema)]
 #[cfg_attr(feature = "generate-ts", derive(ts_rs::TS))]
 #[cfg_attr(
     feature = "generate-ts",
@@ -290,7 +321,7 @@ impl fmt::Display for SelectionChance {
     }
 }
 
-#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema)]
+#[derive(Clone, Debug, Serialize, Deserialize, JsonSchema, ToSchema)]
 #[cfg_attr(feature = "generate-ts", derive(ts_rs::TS))]
 #[cfg_attr(
     feature = "generate-ts",
@@ -311,7 +342,7 @@ impl fmt::Display for InclusionProbabilityResponse {
     }
 }
 
-#[derive(Clone, Serialize, schemars::JsonSchema)]
+#[derive(Clone, Serialize, schemars::JsonSchema, ToSchema)]
 pub struct AllInclusionProbabilitiesResponse {
     pub inclusion_probabilities: Vec<InclusionProbability>,
     pub samples: u64,
@@ -330,7 +361,7 @@ pub struct InclusionProbability {
 
 type Uptime = u8;
 
-#[derive(Clone, Serialize, Deserialize, schemars::JsonSchema)]
+#[derive(Clone, Serialize, Deserialize, schemars::JsonSchema, ToSchema)]
 pub struct MixnodeStatusReportResponse {
     pub mix_id: MixId,
     pub identity: IdentityKey,
@@ -340,22 +371,26 @@ pub struct MixnodeStatusReportResponse {
     pub last_day: Uptime,
 }
 
-#[derive(Clone, Serialize, Deserialize, schemars::JsonSchema)]
+#[derive(Clone, Serialize, Deserialize, schemars::JsonSchema, ToSchema)]
 pub struct GatewayStatusReportResponse {
     pub identity: String,
     pub owner: String,
+    #[schema(value_type = u8)]
     pub most_recent: Uptime,
+    #[schema(value_type = u8)]
     pub last_hour: Uptime,
+    #[schema(value_type = u8)]
     pub last_day: Uptime,
 }
 
-#[derive(Clone, Serialize, Deserialize, schemars::JsonSchema)]
+#[derive(Clone, Serialize, Deserialize, schemars::JsonSchema, ToSchema)]
 pub struct HistoricalUptimeResponse {
     pub date: String,
+    #[schema(value_type = u8)]
     pub uptime: Uptime,
 }
 
-#[derive(Clone, Serialize, Deserialize, schemars::JsonSchema)]
+#[derive(Clone, Serialize, Deserialize, schemars::JsonSchema, ToSchema)]
 pub struct MixnodeUptimeHistoryResponse {
     pub mix_id: MixId,
     pub identity: String,
@@ -363,40 +398,30 @@ pub struct MixnodeUptimeHistoryResponse {
     pub history: Vec<HistoricalUptimeResponse>,
 }
 
-#[derive(Clone, Serialize, Deserialize, schemars::JsonSchema)]
+#[derive(Clone, Serialize, Deserialize, schemars::JsonSchema, ToSchema)]
 pub struct GatewayUptimeHistoryResponse {
     pub identity: String,
     pub owner: String,
     pub history: Vec<HistoricalUptimeResponse>,
 }
 
-/// Because `Coin` is a foreign type, we can't derive `ToSchema` for it. Instead,
-/// use `schema_with` annotation on the fields where it's used:
-///  
-/// https://docs.rs/utoipa-gen/4.3.0/utoipa_gen/derive.ToSchema.html
-fn coin_type() -> Object {
-    ObjectBuilder::new()
-        .schema_type(utoipa::openapi::SchemaType::Object)
-        .property(
-            "denom",
-            ObjectBuilder::new().schema_type(SchemaType::String),
-        )
-        .property(
-            "amount",
-            ObjectBuilder::new().schema_type(SchemaType::Integer),
-        )
-        .build()
+#[derive(ToSchema)]
+#[schema(title = "Coin")]
+pub struct CoinSchema {
+    pub denom: String,
+    #[schema(value_type = String)]
+    pub amount: Uint128,
 }
 
 #[derive(Clone, Serialize, Deserialize, schemars::JsonSchema, ToSchema, ToResponse)]
 pub struct CirculatingSupplyResponse {
-    #[schema(schema_with = coin_type)]
+    #[schema(value_type = CoinSchema)]
     pub total_supply: Coin,
-    #[schema(schema_with = coin_type)]
+    #[schema(value_type = CoinSchema)]
     pub mixmining_reserve: Coin,
-    #[schema(schema_with = coin_type)]
+    #[schema(value_type = CoinSchema)]
     pub vesting_tokens: Coin,
-    #[schema(schema_with = coin_type)]
+    #[schema(value_type = CoinSchema)]
     pub circulating_supply: Coin,
 }
 
@@ -661,7 +686,7 @@ pub struct TestRoute {
     pub layer3: TestNode,
 }
 
-#[derive(Clone, Debug, Serialize, Deserialize, schemars::JsonSchema)]
+#[derive(Clone, Debug, Serialize, Deserialize, schemars::JsonSchema, ToSchema)]
 pub struct PartialTestResult {
     pub monitor_run_id: i64,
     pub timestamp: i64,
