+++ nym-api/nym-api-requests/src/pagination.rs
@@ -3,15 +3,16 @@
 
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
+use utoipa::ToSchema;
 
-#[derive(Serialize, Deserialize, JsonSchema)]
+#[derive(Serialize, Deserialize, JsonSchema, ToSchema)]
 pub struct Pagination {
     pub total: usize,
     pub page: u32,
     pub size: usize,
 }
 
-#[derive(Serialize, Deserialize, JsonSchema)]
+#[derive(Serialize, Deserialize, JsonSchema, ToSchema)]
 pub struct PaginatedResponse<T> {
     pub pagination: Pagination,
     pub data: Vec<T>,
